+++ PizzaTesla/src/main/java/pe/cibertec/edu/service/ProductService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface ProductService {
     List<Product> getProducts();
-    Product getProductById(int id);
+    Product getProductById(Long id);
     Product createProduct(Product product);
     Product updateProduct(Product product);
-    boolean deleteProduct(int id);
-    List<Product> getProductsByCategory(int categoryId);
+    boolean deleteProduct(Long id);
+    List<Product> getProductsByCategory(Long categoryId);
 }
